move both profilers to one file
