+++ bindings/profiler.cc
@@ -16,15 +16,174 @@
 
 #include "v8-profiler.h"
 #include "nan.h"
-#include "time-profiler.h"
-#include "sampling-heap-profiler.h"
+#include <memory>
 
 using namespace v8;
 
-void InitAll(v8::Local<v8::Object> target) {
-  Nan::HandleScope scope;
-  SamplingHeapProfilerInit(target);
-  TimeProfilerInit(target);
+// Sampling Heap Profiler
+
+Local<Value> TranslateAllocationProfile(AllocationProfile::Node* node) {
+  Local<Object> js_node = Nan::New<Object>();
+  js_node->Set(Nan::New<String>("name").ToLocalChecked(),
+    node->name);
+  js_node->Set(Nan::New<String>("scriptName").ToLocalChecked(),
+    node->script_name);
+  js_node->Set(Nan::New<String>("scriptId").ToLocalChecked(),
+    Nan::New<Integer>(node->script_id));
+  js_node->Set(Nan::New<String>("lineNumber").ToLocalChecked(),
+    Nan::New<Integer>(node->line_number));
+  js_node->Set(Nan::New<String>("columnNumber").ToLocalChecked(),
+    Nan::New<Integer>(node->column_number));
+  Local<Array> children = Nan::New<Array>(node->children.size());
+  for (size_t i = 0; i < node->children.size(); i++) {
+    children->Set(i, TranslateAllocationProfile(node->children[i]));
+  }
+  js_node->Set(Nan::New<String>("children").ToLocalChecked(),
+    children);
+  Local<Array> allocations = Nan::New<Array>(node->allocations.size());
+  for (size_t i = 0; i < node->allocations.size(); i++) {
+    AllocationProfile::Allocation alloc = node->allocations[i];
+    Local<Object> js_alloc = Nan::New<Object>();
+    js_alloc->Set(Nan::New<String>("sizeBytes").ToLocalChecked(),
+      Nan::New<Number>(alloc.size));
+    js_alloc->Set(Nan::New<String>("count").ToLocalChecked(),
+      Nan::New<Number>(alloc.count));
+    allocations->Set(i, js_alloc);
+  }
+  js_node->Set(Nan::New<String>("allocations").ToLocalChecked(),
+    allocations);
+  return js_node;
 }
 
-NODE_MODULE(profiler, InitAll)
+NAN_METHOD(StartSamplingHeapProfiler) {
+  if (info.Length() == 2) {
+    if (!info[0]->IsUint32()) {
+      return Nan::ThrowTypeError("First argument type must be uint32.");
+    }
+    if (!info[1]->IsNumber()) {
+      return Nan::ThrowTypeError("First argument type must be Integer.");
+    }
+
+#if NODE_MODULE_VERSION > NODE_8_0_MODULE_VERSION
+    uint64_t sample_interval = info[0].As<Integer>()->Value();
+    int stack_depth = info[1].As<Integer>()->Value();
+#else
+    uint64_t sample_interval = info[0].As<Integer>()->Uint32Value();
+    int stack_depth = info[1].As<Integer>()->IntegerValue();
+#endif
+
+    info.GetIsolate()->GetHeapProfiler()->
+      StartSamplingHeapProfiler(sample_interval, stack_depth);
+  } else {
+    info.GetIsolate()->GetHeapProfiler()->StartSamplingHeapProfiler();
+  }
+}
+
+NAN_METHOD(StopSamplingHeapProfiler) {
+  info.GetIsolate()->GetHeapProfiler()->StopSamplingHeapProfiler();
+}
+
+NAN_METHOD(GetAllocationProfile) {
+  std::unique_ptr<v8::AllocationProfile> profile(
+    info.GetIsolate()->GetHeapProfiler()->GetAllocationProfile());
+  AllocationProfile::Node* root = profile->GetRootNode();
+  info.GetReturnValue().Set(TranslateAllocationProfile(root));
+}
+
+
+// Time profiler
+
+#if NODE_MODULE_VERSION > NODE_8_0_MODULE_VERSION
+// This profiler exists for the lifetime of the program. Not calling 
+// CpuProfiler::Dispose() is intentional.
+CpuProfiler* cpuProfiler = CpuProfiler::New(v8::Isolate::GetCurrent());
+#else
+CpuProfiler* cpuProfiler = v8::Isolate::GetCurrent()->GetCpuProfiler();
+#endif
+
+
+Local<Value> TranslateTimeProfileNode(const CpuProfileNode* node) {
+  Local<Object> js_node = Nan::New<Object>();
+  js_node->Set(Nan::New<String>("name").ToLocalChecked(),
+    node->GetFunctionName());
+  js_node->Set(Nan::New<String>("scriptName").ToLocalChecked(),
+    node->GetScriptResourceName());
+  js_node->Set(Nan::New<String>("scriptId").ToLocalChecked(),
+    Nan::New<Integer>(node->GetScriptId()));
+  js_node->Set(Nan::New<String>("lineNumber").ToLocalChecked(),
+    Nan::New<Integer>(node->GetLineNumber()));
+  js_node->Set(Nan::New<String>("columnNumber").ToLocalChecked(),
+    Nan::New<Integer>(node->GetColumnNumber()));
+  js_node->Set(Nan::New<String>("hitCount").ToLocalChecked(),
+    Nan::New<Integer>(node->GetHitCount()));
+  int32_t count = node->GetChildrenCount();
+  Local<Array> children = Nan::New<Array>(count);
+  for (int32_t i = 0; i < count; i++) {
+    children->Set(i, TranslateTimeProfileNode(node->GetChild(i)));
+  }
+  js_node->Set(Nan::New<String>("children").ToLocalChecked(),
+    children);
+  return js_node;
+}
+
+Local<Value> TranslateTimeProfile(const CpuProfile* profile) {
+  Local<Object> js_profile = Nan::New<Object>();
+  js_profile->Set(Nan::New<String>("title").ToLocalChecked(),
+    profile->GetTitle());
+  js_profile->Set(Nan::New<String>("topDownRoot").ToLocalChecked(),
+    TranslateTimeProfileNode(profile->GetTopDownRoot()));
+  js_profile->Set(Nan::New<String>("startTime").ToLocalChecked(),
+    Nan::New<Number>(profile->GetStartTime()));
+  js_profile->Set(Nan::New<String>("endTime").ToLocalChecked(),
+    Nan::New<Number>(profile->GetEndTime()));
+  return js_profile;
+}
+
+NAN_METHOD(StartProfiling) {
+  Local<String> name = info[0].As<String>();
+
+  // Sample counts and timestamps are not used, so we do not need to record
+  // samples.
+  cpuProfiler->StartProfiling(name, false);
+}
+
+NAN_METHOD(StopProfiling) {
+  Local<String> name = info[0].As<String>();
+  CpuProfile* profile =
+    cpuProfiler->StopProfiling(name);
+  Local<Value> translated_profile = TranslateTimeProfile(profile);
+  profile->Delete();
+  info.GetReturnValue().Set(translated_profile);
+}
+
+NAN_METHOD(SetSamplingInterval) {
+#if NODE_MODULE_VERSION > NODE_8_0_MODULE_VERSION
+  int us = info[0].As<Integer>()->Value();
+#else
+  int us = info[0].As<Integer>()->IntegerValue();
+#endif
+  cpuProfiler->SetSamplingInterval(us);
+}
+
+
+NAN_MODULE_INIT(InitAll) {
+  Local<Object> timeProfiler = Nan::New<Object>();
+  Nan::Set(timeProfiler, Nan::New("startProfiling").ToLocalChecked(),
+    Nan::GetFunction(Nan::New<FunctionTemplate>(StartProfiling)).ToLocalChecked());
+  Nan::Set(timeProfiler, Nan::New("stopProfiling").ToLocalChecked(),
+    Nan::GetFunction(Nan::New<FunctionTemplate>(StopProfiling)).ToLocalChecked());
+  Nan::Set(timeProfiler, Nan::New("setSamplingInterval").ToLocalChecked(),
+    Nan::GetFunction(Nan::New<FunctionTemplate>(SetSamplingInterval)).ToLocalChecked());
+  target->Set(Nan::New<String>("timeProfiler").ToLocalChecked(), timeProfiler);
+
+  Local<Object> heapProfiler = Nan::New<Object>();
+  Nan::Set(heapProfiler, Nan::New("startSamplingHeapProfiler").ToLocalChecked(),
+    Nan::GetFunction(Nan::New<FunctionTemplate>(StartSamplingHeapProfiler)).ToLocalChecked());
+  Nan::Set(heapProfiler, Nan::New("stopSamplingHeapProfiler").ToLocalChecked(),
+    Nan::GetFunction(Nan::New<FunctionTemplate>(StopSamplingHeapProfiler)).ToLocalChecked());
+  Nan::Set(heapProfiler, Nan::New("getAllocationProfile").ToLocalChecked(),
+    Nan::GetFunction(Nan::New<FunctionTemplate>(GetAllocationProfile)).ToLocalChecked());
+  target->Set(Nan::New<String>("heapProfiler").ToLocalChecked(), heapProfiler);
+}
+
+NODE_MODULE(profiler, InitAll);
