--- bindings/sampling-heap-profiler.cc
@@ -1,103 +1,0 @@
-/**
- * Copyright 2015 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "v8-profiler.h"
-#include "nan.h"
-#include "sampling-heap-profiler.h"
-#include <memory>
-
-using namespace v8;
-
-Local<Value> TranslateAllocationProfile(AllocationProfile::Node* node) {
-  Local<Object> js_node = Nan::New<Object>();
-  js_node->Set(Nan::New<String>("name").ToLocalChecked(),
-    node->name);
-  js_node->Set(Nan::New<String>("scriptName").ToLocalChecked(),
-    node->script_name);
-  js_node->Set(Nan::New<String>("scriptId").ToLocalChecked(),
-    Nan::New<Integer>(node->script_id));
-  js_node->Set(Nan::New<String>("lineNumber").ToLocalChecked(),
-    Nan::New<Integer>(node->line_number));
-  js_node->Set(Nan::New<String>("columnNumber").ToLocalChecked(),
-    Nan::New<Integer>(node->column_number));
-  Local<Array> children = Nan::New<Array>(node->children.size());
-  for (size_t i = 0; i < node->children.size(); i++) {
-    children->Set(i, TranslateAllocationProfile(node->children[i]));
-  }
-  js_node->Set(Nan::New<String>("children").ToLocalChecked(),
-    children);
-  Local<Array> allocations = Nan::New<Array>(node->allocations.size());
-  for (size_t i = 0; i < node->allocations.size(); i++) {
-    AllocationProfile::Allocation alloc = node->allocations[i];
-    Local<Object> js_alloc = Nan::New<Object>();
-    js_alloc->Set(Nan::New<String>("sizeBytes").ToLocalChecked(),
-      Nan::New<Number>(alloc.size));
-    js_alloc->Set(Nan::New<String>("count").ToLocalChecked(),
-      Nan::New<Number>(alloc.count));
-    allocations->Set(i, js_alloc);
-  }
-  js_node->Set(Nan::New<String>("allocations").ToLocalChecked(),
-    allocations);
-  return js_node;
-}
-
-NAN_METHOD(StartSamplingHeapProfiler) {
-  if (info.Length() == 2) {
-    if (!info[0]->IsUint32()) {
-      return Nan::ThrowTypeError("First argument type must be uint32.");
-    }
-    if (!info[1]->IsNumber()) {
-      return Nan::ThrowTypeError("First argument type must be Integer.");
-    }
-
-#if NODE_MODULE_VERSION > NODE_8_0_MODULE_VERSION
-    uint64_t sample_interval = info[0].As<Integer>()->Value();
-    int stack_depth = info[1].As<Integer>()->Value();
-#else
-    uint64_t sample_interval = info[0].As<Integer>()->Uint32Value();
-    int stack_depth = info[1].As<Integer>()->IntegerValue();
-#endif
-
-    info.GetIsolate()->GetHeapProfiler()->
-      StartSamplingHeapProfiler(sample_interval, stack_depth);
-  } else {
-    info.GetIsolate()->GetHeapProfiler()->StartSamplingHeapProfiler();
-  }
-}
-
-NAN_METHOD(StopSamplingHeapProfiler) {
-  info.GetIsolate()->GetHeapProfiler()->StopSamplingHeapProfiler();
-}
-
-NAN_METHOD(GetAllocationProfile) {
-  std::unique_ptr<v8::AllocationProfile> profile(
-    info.GetIsolate()->GetHeapProfiler()->GetAllocationProfile());
-  AllocationProfile::Node* root = profile->GetRootNode();
-  info.GetReturnValue().Set(TranslateAllocationProfile(root));
-}
-
-void SamplingHeapProfilerInit(Local<Object> target) {
-  Local<Object> heapProfiler = Nan::New<Object>();
-
-  Nan::Set(heapProfiler, Nan::New("startSamplingHeapProfiler").ToLocalChecked(),
-    Nan::GetFunction(Nan::New<FunctionTemplate>(StartSamplingHeapProfiler)).ToLocalChecked());
-  Nan::Set(heapProfiler, Nan::New("stopSamplingHeapProfiler").ToLocalChecked(),
-    Nan::GetFunction(Nan::New<FunctionTemplate>(StopSamplingHeapProfiler)).ToLocalChecked());
-  Nan::Set(heapProfiler, Nan::New("getAllocationProfile").ToLocalChecked(),
-    Nan::GetFunction(Nan::New<FunctionTemplate>(GetAllocationProfile)).ToLocalChecked());
-
-  target->Set(Nan::New<String>("heapProfiler").ToLocalChecked(), heapProfiler);
-}
