--- bindings/sampling-heap-profiler.h
@@ -1,30 +1,0 @@
-/**
- * Copyright 2018 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- 
-#ifndef SAMPLING_HEAP_PROFILER
-#define SAMPLING_HEAP_PROFILER
-
-#include "v8-profiler.h"
-#include "nan.h"
-
-using namespace v8;
-
-NAN_METHOD(StartSamplingHeapProfiler);
-NAN_METHOD(StopSamplingHeapProfiler);
-NAN_METHOD(GetAllocationProfile);
-void SamplingHeapProfilerInit(Local<Object> target);
-
-#endif
